+++ instrumentation/opentelemetry-instrumentation-loguru/tests/test_loguru.py
@@ -1,0 +1,66 @@
+# Copyright The OpenTelemetry Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Optional
+
+import pytest
+
+from opentelemetry.instrumentation.loguru import (  # pylint: disable=no-name-in-module
+    LoguruInstrumentor,
+)
+from opentelemetry.test.test_base import TestBase
+from opentelemetry.trace import ProxyTracer, get_tracer
+
+
+class FakeTracerProvider:
+    def get_tracer(  # pylint: disable=no-self-use
+        self,
+        instrumenting_module_name: str,
+        instrumenting_library_version: Optional[str] = None,
+        schema_url: Optional[str] = None,
+    ) -> ProxyTracer:
+        return ProxyTracer(
+            instrumenting_module_name,
+            instrumenting_library_version,
+            schema_url,
+        )
+
+
+class TestLoguruInstrumentorProxyTracerProvider(TestBase):
+    @pytest.fixture(autouse=True)
+    def inject_fixtures(self, caplog):
+        self.caplog = caplog  # pylint: disable=attribute-defined-outside-init
+
+    def setUp(self):
+        super().setUp()
+        LoguruInstrumentor().instrument(tracer_provider=FakeTracerProvider())
+
+    def tearDown(self):
+        super().tearDown()
+        LoguruInstrumentor().uninstrument()
+
+
+class TestLoguruInstrumentor(TestBase):
+    @pytest.fixture(autouse=True)
+    def inject_fixtures(self, caplog):
+        self.caplog = caplog  # pylint: disable=attribute-defined-outside-init
+
+    def setUp(self):
+        super().setUp()
+        LoguruInstrumentor().instrument()
+        self.tracer = get_tracer(__name__)
+
+    def tearDown(self):
+        super().tearDown()
+        LoguruInstrumentor().uninstrument()